requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
Authlib==1.6.5
beartype==0.22.8
beautifulsoup4==4.14.3
bota==4.0.155
botasaurus==4.0.96
botasaurus-humancursor==4.0.83
botasaurus_api==4.0.10
botasaurus_driver==4.0.92
botasaurus_proxy_authentication==1.0.16
botasaurus_requests==4.0.38
brotli==1.2.0
bs4==0.0.2
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
close_chrome==4.0.40
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
cyclopts==4.3.0
DateTime==6.0
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.3.1
diskcache==5.6.3
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastmcp==2.13.3
gevent==25.9.1
geventhttpclient==2.3.5
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
javascript_fixes==1.1.29
jedi==0.19.2
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
lxml==6.0.2
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mcp==1.22.0
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.3.5
openapi-pydantic==0.5.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathable==0.4.4
pathvalidate==3.3.1
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pyperclip==1.11.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
PyVirtualDisplay==3.0
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
ScraperFC==3.4.0
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.45
sqlmodel==0.0.27
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
tavily-python==0.7.21
tiktoken==0.12.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
